backend/pyproject.toml
[tool.poetry]
name = "sales-forecasting-backend"
version = "1.0.0"
description = "Production-grade Sales Forecasting & Marketing Mix Modeling API"
authors = ["Sales Forecasting Team"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"

# Web Framework
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
python-multipart = "^0.0.6"

# GraphQL
strawberry-graphql = {extras = ["fastapi"], version = "^0.217.0"}

# Database
sqlalchemy = {extras = ["asyncio"], version = "^2.0.25"}
asyncpg = "^0.29.0"
alembic = "^1.13.1"
psycopg2-binary = "^2.9.9"

# Cache
redis = "^5.0.1"

# Authentication & Security
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
bcrypt = "^4.1.2"

# Task Queue
celery = {extras = ["redis"], version = "^5.3.6"}

# ML & Data Science
numpy = "^1.26.3"
pandas = "^2.1.4"
scikit-learn = "^1.4.0"
scipy = "^1.12.0"
statsmodels = "^0.14.1"

# Forecasting
prophet = "^1.1.5"

# Bayesian MMM
pymc = "^5.10.3"
pymc-marketing = "^0.4.0"
arviz = "^0.17.0"

# MLOps
mlflow = "^2.10.0"
dvc = {extras = ["s3"], version = "^3.38.1"}

# Data Validation
great-expectations = "^0.18.8"
pydantic = "^2.5.3"
pydantic-settings = "^2.1.0"

# Storage
boto3 = "^1.34.25"

# Utilities
python-dateutil = "^2.8.2"
httpx = "^0.26.0"
tenacity = "^8.2.3"
structlog = "^24.1.0"
orjson = "^3.9.12"

# Excel/CSV handling
openpyxl = "^3.1.2"
xlsxwriter = "^3.1.9"

# Optimization
cvxpy = "^1.4.2"

# ============================================
# ADVANCED ANALYTICS & ML ADDITIONS
# ============================================

# Causal Inference & Attribution
causalml = "^0.15.0"
dowhy = "^0.11.0"
causalnex = "^0.12.0"
pgmpy = "^0.1.24"

# Deep Learning & Neural Networks
torch = "^2.1.0"
pytorch-lightning = "^2.1.0"
pytorch-forecasting = "^1.0.0"
neuralforecast = "^1.6.0"
gluonts = "^0.14.0"
darts = "^0.27.0"

# AutoML & Explainability
optuna = "^3.5.0"
shap = "^0.44.0"
lime = "^0.2.0"
mapie = "^0.8.0"

# Streaming & Real-Time
aiokafka = "^0.10.0"
websockets = "^12.0"

# Online Learning & Drift Detection
river = "^0.21.0"
alibi-detect = "^0.12.0"

# Privacy-Preserving Analytics
pydp = "^1.1.5"
opendp = "^0.9.0"

# Enterprise Authentication
python3-saml = "^1.16.0"
authlib = "^1.3.0"

# Additional Data Connectors
snowflake-connector-python = "^3.6.0"
google-cloud-bigquery = "^3.14.0"
databricks-sql-connector = "^3.0.0"
google-ads = "^23.1.0"
facebook-business = "^19.0.0"
salesforce-api = "^0.1.0"
simple-salesforce = "^1.12.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-asyncio = "^0.23.3"
pytest-cov = "^4.1.0"
httpx = "^0.26.0"
mypy = "^1.8.0"
ruff = "^0.1.14"
black = "^24.1.1"
pre-commit = "^3.6.0"
factory-boy = "^3.3.0"
faker = "^22.2.0"
aiosqlite = "^0.19.0"

[tool.poetry.group.robyn]
optional = true

[tool.poetry.group.robyn.dependencies]
rpy2 = "^3.5.14"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py311"
line-length = 100
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.isort]
known-first-party = ["app"]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --cov=app --cov-report=term-missing"
